Add rudimentary aspect ratio handling
In general, neither the rendering target nor the shown part of the
tone net is square.  However, the coordinate system used for rendering
is set up such that nodes coincide exactly with coordinates consisting
of integral components.  Thus, for square nodes, we need to scale either
width or height to accomodate aspect ratios of both rendering target and
coordinate system.

Signed-off-by: Maximilian Marx <mmarx@wh2.tu-dresden.de>